+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change history for ui-linked-data
 
 ## 3.0.0 (IN PROGRESS)
+* Fix default profile type persistence across edit form and profile settings. Fixes [UILD-820].
+
+[UILD-820]:https://folio-org.atlassian.net/browse/UILD-820
 
 ## 2.0.3 (2026-05-19)
 * Fix Reset search option is missing after using Advanced Search. Fixes [UILD-812].